Start work on caching coastline
HelenOS headers and libraries are copied to the build directory, thus
allowing to rebuild HelenOS for other architectures. The big advantage
is that after first coastline build you do not need to care whether
HelenOS is configured for the rigth architecture.

This is a first commit from the transition to caching scheme, some
packages may no longer work. Will be fixed later.